fix oos handling in metrics
